requirements.txt
# biblioteki

flask
requests
PyYAML
SQLAlchemy
openpyxl
flask-login
